Move ASP.NET Core hosting extensions to own assembly/package and make test service use it
