+++ samples/MvcTestService/MvcTestService.csproj
@@ -5,15 +5,15 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\DasMulli.Win32.ServiceUtils\DasMulli.Win32.ServiceUtils.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.All" Version="2.0.3" />
   </ItemGroup>
 
   <ItemGroup>
     <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\DasMulli.AspNetCore.Hosting.WindowsServices\DasMulli.AspNetCore.Hosting.WindowsServices.csproj" />
   </ItemGroup>
 
 </Project>
